Movendo HTTPEntryPointImpl para um pacote separado
